+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+-   Fixed issue with install path not being used properly when switching empire location (@Vinnybod)
 -   Lock nim version in the install script (@Vinnybod)
 
 ## [5.2.0] - 2023-03-31